11.1.4
Installed express.js in the root by typing npm init -y.   The -y adds auto yes responses to the npm init questionnaire which seemed fine in this case.  Then we installed express.js npm i express.  "i" is a shorthand for install.  Finally we required express in our server.js and then made a call that looked like this
const app = express();.

Afterwards we did this
app.listen(3001, () => {
  console.log(`API server now on port 3001!`);
});

Which used the variable for the call  and opened a port with .listen().  Now I have port 3001 opened on my terminal and am awaiting the next steps.