11.2.3
Introduction to the idea of POST which means the client (person using the site) is given permission to add to the database within the scope of what we've given them permission to do with POST.  This is the exact opposite of GET.

app.post("/api/animals", (req, res) => {
  // req.body is where our incoming content will be
  console.log(req.body);
  res.json(req.body);
});
we are allowing people to tweak with the animals api.  req.body is an example of what the file is that they will be sending.  Here we console log it just to see it before it leaves and use the res.json to see it after it gets there.