11.2.4
Set up Insomnia to test the GET.  Worth noting we had to go to the right to select Create New Request to get the screen to follow the module.  Once in we then on the left selected new request and then our typed our http://localhost:3001/api/animals into the top.  Then we chose the query tab to test our GETs to make sure we can sort things by name.  It is worth noting that I never got the params using id to work with query though you can just type it up top and everything is fine..  Then we changed it from GET to POST and selected JSON.  Then we added an object matching the format of all the other animals and hit send.  This worked with a 200ok but sent back undefined because the object is not defined as json yet in our server.