11.2.6
Added a
app.use(express.urlencoded({ extended: true }));
to first convert the key/values so that it can be used by the req.body object.  Then the extended true warns it that there is at least one array in there and to go deeper and check for them so it can parse this data correctly.  It also looks like if you take the array out of the object you can still keep this code in.

Finally we use
app.use(express.json());
to convert the data into a usabel json format by the req.body object.