11.3.4
Got the index.html to display by adding the this app.get to the server.js

app.get("/", (req, res) => {
  res.sendFile(path.join(__dirname, "./public/index.html"));
});
this uses res.sedFile to write the page and uses path.join to find it using __dirname to represent where we are and ,"./public/index.html" to represent where we want to go from with server.js being the stat point.  Then we discover the page only pushed the HTML and not the CSS or the JS. To fix this we use another express function

app.use(express.static("public"));
express.static basically tells it to include this in all deploys of this page.  In this case we are targeting the "public" folder which contains all of the front end information.