11.3.5
Added a Fetch to the handleAnimalFormSubmit that honestly wasn't explained very well.

const animalObject = { name, species, diet, personalityTraits };
this is the info from the top part that was written that contains all of the answers.  Then we take this object variable and.

  fetch("/api/animals", {
    method: "POST",
    headers: {
      Accept: "application/json",
      "Content-Type": "application/json",
    },
this goes to our server.js and looks for our /api/animals POST and let it know that it is json time.

    body: JSON.stringify(animalObject),
then we tell it to use the animalObject we just created with all the server finctions and apply it to the body.
  })
    .then((response) => {
      if (response.ok) {
        return response.json();
      }
      alert("Error: " + response.statusText);
    })
    .then((postResponse) => {
      console.log(postResponse);
      alert("Thank you for adding an animal!");
    });
};