11.3.6
Added  3 new routs to get the page working.  The previoulsy set up express.static referencing the public folder applies to these ensuring they get there CSS and JS files attached when deployed.

app.get("/animals", (req, res) => {
  res.sendFile(path.join(__dirname, "./public/animals.html"));
});
Gets the animals html page and displays it.

app.get("/zookeepers", (req, res) => {
  res.sendFile(path.join(__dirname, "./public/zookeepers.html"));
});
Finds the zoopkeepers page and siplays it.  Fun side note it first displays an error because it's JS already has a Fetch cooked in that fails upon loading.  Not mentioned at all in the module of course.

app.get("*", (req, res) => {
  res.sendFile(path.join(__dirname, "./public/index.html"));
});
This is intended to be a catch all for anything that fails will redirect the client to the index.HTML