11.3.7
added a new Fetch to the animals.js to display go get the info from animals.json.

fetch(queryUrl)
    .then((response) => {
      if (!response.ok) {
        return alert("Error: " + response.statusText);
      }
      return response.json();
    })
    .then((animalData) => {
      console.log(animalData);
      printResults(animalData);
    });
};
queryUrl looks like "/api/animals/?"  This then goes to our server.json and gets it's info.  If it's response is ok it json's it and then uses the printResults() already built for us on this page to display it to the page.