11.4.3
we moved all the functions out of server.js in the root and placed them in the newly created animals.js in the lib folder.  Then we required fs and path up top and exported all of the functions using module.exports= {} at the bottom.  None of the functions works yet as we are not talking to this file.