11.4.4
Added a ton of files.  We made an apiRoutes Folder in the routes folder  that we then added animals.js to.  We moved all of our app.get('/api/animals', (req, res) => {} into here imported the functions from lib/animals and the api object from data/animals.  It is worth noting to get to api/animals we had to jump up to folder levels using ../../data/animals.  Then we were introduced to ROUTER which we put at the top of our code like this const router = require('express').Router(); and then had to change every app.get to router.get.  At the bottom we did a module.export=router; to send it out.

Then we went to to set up our html by creating an htmlRoutes folder in routes and then added the index.js file.  Then we went to server.s and grabbed all the app.get that DID NOT say api which means they actually apply to the html and moved them in.  Following the same process we brought in ROUTER const router = require("express").Router(); changed every app.get to router.get and used module.exports = router to send them out.  We also needed to require path to make them wrok seeing as how the existing code was already leaning on path.

next we went back into the apiRoutes folder and created a new index.js in there.  In here we yet again required ROUTER but then also required  a the file animalRoutes that we share a folder with.  The weird thing here is to access it we actually backed all the way out to the file we share (apiRoutes) and then went back in like this const animalRoutes = require('../apiRoutes/animalRoutes');.  They claim we did this as a future proofing thing for additional routes and we can just simply reference the apiRoutes file.    Then we used router.use(animalRoutes); making index.js the central hub for all api routing functions.

Finally after all of that we went back into the now barren server.js and added
const apiRoutes = require('./routes/apiRoutes');
const htmlRoutes = require('./routes/htmlRoutes');
up top to reference the index.js files in those locations.  Finally we added

app.use('/api', apiRoutes);
app.use('/', htmlRoutes);
to glue it all together.