11.4.5
Created a test for the functions in animals.js.  This was a blatant copy pasta so not much teaching here.  The one issue that came up was that when we ran the test it did add to the back end.  To get rid of that we added jest.mock("fs") to make sure that the test fakes it?