11.4.6
More copy pasta setting up the zookeepers.json in the data folder being the home for the back end zookeeper object.
 Then in the lib folder we created the zookeepers.js and added all the functions for the program.  All of which is very similar to animals.js.
Then went into the __tests__ folder and created tests for the newly created functions in the zookeepers.test.js files and watched them pass.
Then we went into the apiRoutes folder within the routes folder to then add all the zooKeeper api routes in the animalRoutes.js file.  Again this is very similar to how the animalsRoutes.js looks we just had to change the names.
Finally we stayed in the same apiRoutes folder and updated the index.js to include --router.use(require("./zookeeperRoutes"));-- and tested it in the browser which worked.  I then commented that out and wrote it like this
const zookeeperRoutes = require("../apiRoutes/zookeeperRoutes");
router.use(zookeeperRoutes);
to remain consistent with the way we wrote it before and this also worked.